+++ haste-preset-yoshi/src/commands/build.js
@@ -51,10 +51,10 @@ module.exports = runner.command(async tasks => {
 
   await Promise.all([
     clean({pattern: `{dist,target}/*`}),
-    wixUpdateNodeVersion({}, {title: 'update-node-version'}),
-    migrateScopePackages({}, {title: 'scope-packages-migration'}),
-    migrateBowerArtifactory({}, {title: 'migrate-bower-artifactory'}),
-    wixDepCheck({}, {title: 'dep-check'})
+    wixUpdateNodeVersion({}, {title: 'update-node-version', log: false}),
+    migrateScopePackages({}, {title: 'scope-packages-migration', log: false}),
+    migrateBowerArtifactory({}, {title: 'migrate-bower-artifactory', log: false}),
+    wixDepCheck({}, {title: 'dep-check', log: false})
   ]);
 
   await Promise.all([
@@ -64,11 +64,11 @@ module.exports = runner.command(async tasks => {
       `${globs.base()}/assets/**/*`,
       `${globs.base()}/**/*.{ejs,html,vm}`,
       `${globs.base()}/**/*.{css,json,d.ts}`,
-    ], target: 'dist'}, {title: 'copy-server-assets'}),
+    ], target: 'dist'}, {title: 'copy-server-assets', log: false}),
     copy({pattern: [
       `${globs.assetsLegacyBase()}/assets/**/*`,
       `${globs.assetsLegacyBase()}/**/*.{ejs,html,vm}`,
-    ], target: 'dist/statics'}, {title: 'copy-static-assets-legacy'}),
+    ], target: 'dist/statics'}, {title: 'copy-static-assets-legacy', log: false}),
     copy({
       pattern: [
         `assets/**/*`,
@@ -76,13 +76,13 @@ module.exports = runner.command(async tasks => {
       ],
       source: globs.assetsBase(),
       target: 'dist/statics'
-    }, {title: 'copy-static-assets'}),
+    }, {title: 'copy-static-assets', log: false}),
     bundle(),
-    wixPetriSpecs({config: petriSpecsConfig()}, {title: 'petri-specs'}),
+    wixPetriSpecs({config: petriSpecsConfig()}, {title: 'petri-specs', log: false}),
     wixMavenStatics({
       clientProjectName: clientProjectName(),
       staticsDir: clientFilesPath()
-    }, {title: 'maven-statics'})
+    }, {title: 'maven-statics', log: false})
   ]);
 
   function bundle() {